refactor: add Status and DocumentStatus components
`Editor/Status` handles the status display in the menubar.
`Editor/DocumentStatus` handles the document messages on conflict etc.

Also move all files that are only used within it into the `Components/Editor` dir.
