+++ src/components/Editor/AvatarWrapper.vue
